move beholder to base cell class
